Moving nextDigestAt from individual activeTopicSubscriptions to the User record. We send digests at a user level, so it no longer makes sense to track digests individually by topic
