+++ functions/src/notifications/updateUserNotificationFrequency.ts
@@ -1,6 +1,6 @@
 const functions = require("firebase-functions")
 import * as admin from "firebase-admin"
-import { Timestamp } from "../firebase"
+import { getNextDigestAt } from "./helpers";
 
 export const updateUserNotificationFrequency = functions.firestore
   .document("profiles/{userId}")
@@ -38,52 +38,11 @@ export const updateUserNotificationFrequency = functions.firestore
       // Update user document in the 'users' collection
       await admin.firestore().collection("users").doc(userId).set(
         {
-          notificationFrequency: notificationFrequency
+          notificationFrequency: notificationFrequency,
+          nextDigestAt: getNextDigestAt(notificationFrequency)
         },
         { merge: true }
       )
-
-      // Update all documents in the 'activeTopicSubscriptions' sub-collection with the new nextDigestAt value
-      const now = Timestamp.now()
-      let nextDigestAt: admin.firestore.Timestamp | null
-      switch (notificationFrequency) {
-        case "Daily":
-          nextDigestAt = Timestamp.fromMillis(
-            now.toMillis() + 24 * 60 * 60 * 1000
-          )
-          break
-        case "Weekly":
-          nextDigestAt = Timestamp.fromMillis(
-            now.toMillis() + 7 * 24 * 60 * 60 * 1000
-          )
-          break
-        case "Monthly":
-          const monthAhead = new Date(now.toDate())
-          monthAhead.setMonth(monthAhead.getMonth() + 1)
-          nextDigestAt = Timestamp.fromDate(monthAhead)
-          break
-        case "None":
-          nextDigestAt = null
-          break
-        default:
-          console.error(
-            `Unknown notification frequency: ${notificationFrequency}`
-          )
-          break
-      }
-
-      const subscriptionDocs = await admin
-        .firestore()
-        .collection("users")
-        .doc(userId)
-        .collection("activeTopicSubscriptions")
-        .get()
-      const batch = admin.firestore().batch()
-      subscriptionDocs.docs.forEach((doc: { ref: any }) => {
-        batch.update(doc.ref, { nextDigestAt: nextDigestAt })
-      })
-
-      await batch.commit()
 
       return null
     }
