+++ functions/src/subscriptions/subscribeToBillTopic.ts
@@ -2,7 +2,6 @@ import { Database } from "../types"
 import { UserRecord } from "firebase-admin/auth"
 import { TopicSubscription } from "./types"
 import { addTopicSubscription } from "./addTopicSubscription"
-import { Timestamp } from "../firebase"
 
 export const subscribeToBillTopic = async ({
   user,
@@ -20,8 +19,7 @@ export const subscribeToBillTopic = async ({
     topicName,
     uid,
     type: "bill",
-    billLookup,
-    nextDigestAt: Timestamp.fromDate(new Date())
+    billLookup
   }
 
   await addTopicSubscription({ user, subscriptionData, db })
