+++ functions/src/subscriptions/types.ts
@@ -11,5 +11,4 @@ export interface TopicSubscription {
     profileId: string
     fullName: string
   }
-  nextDigestAt: any
 }
